main task done

+++ lab2/src/main/java/org/testing/math/log/AbsLogFunc.java
@@ -5,7 +5,7 @@ import org.testing.math.CSVInterface;
 public abstract class AbsLogFunc implements CSVInterface {
     LogBaseN ln;
 
-    public abstract Double solveLog(double x, long base, double eps);
+    public abstract Double solveLog(double x, double base, double eps);
     public abstract void setLn(LogBaseN ln);
 
 }
